+++ src/lib.rs
@@ -1,3 +1,11 @@
-// src/lib.rs
-
 pub mod filter;
+pub mod git;
+pub mod path;
+pub mod template;
+pub mod token;
+
+pub use filter::should_include_file;
+pub use git::get_git_diff;
+pub use path::{traverse_directory, label};
+pub use template::{handlebars_setup, render_template, extract_undefined_variables};
+pub use token::count_tokens;
